requirements.txt
fastapi
uvicorn
psycopg2-binary
python-decouple==3.5